Fixed stdout logging whend passed as CMake argument. Added script to run benchmark with different memory settings. Changed some types to conform libvirt api parameter types. Fixed KiB/MiB issue in benchmark. Improved some error messages by adding hypervisor information. Further improvements on error messages are about to come.

+++ src/hypervisor.hpp
@@ -12,7 +12,7 @@ class Hypervisor
 {
 public:
 	virtual ~Hypervisor() {};
-	virtual void start(const std::string &vm_name, size_t vcpus, size_t memory) = 0;
+	virtual void start(const std::string &vm_name, unsigned int vcpus, unsigned long memory) = 0;
 	virtual void stop(const std::string &vm_name) = 0;
 	virtual void migrate(const std::string &vm_name, const std::string &dest_hostname, bool live_migration) = 0;
 };
